remove try_start loop
If started as a daemon, the parent mongodb process waits for the child
to be ready to accept connections before exiting. The rocketchat-mongo
app now takes care of the logs. The "after" keys make sure both
rocketchat-{server,mongo} starts after mongodb is ready for connections
and the log file is there to be read. Since the service type (or daemon
type) is "forking", units that depend on this, are not going to start
unless the parent process exits.