ci: don't calculate greater than as it is all consuming and can fail for lack of context
If there are two version strings, and we perform a greater than, in this
way, since every next step doesn't know about the last one, it can
result in a false positive.

Instead, if we test if A is less than B, for the three steps, if the
previous did not cause a return, means we are still ok. Therefore this
current step is all we need.